requirements.txt
aiosmtplib==1.1.6
alembic==1.11.1
algoliasearch==3.0.0
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
attrs==23.1.0
bcrypt==4.0.1
boto3==1.28.1
botocore==1.31.1
camel-converter==3.0.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
cryptography==36.0.2
Deprecated==1.2.13
ecdsa==0.18.0
et-xmlfile==1.1.0
fastapi==0.100.0
filelock==3.12.2
h11==0.14.0
hiredis==2.2.3
httpcore==0.16.3
httpx==0.23.3
idna==3.4
jmespath==1.0.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
loguru==0.7.0
Mako==1.2.4
mangum==0.17.0
MarkupSafe==2.1.3
meilisearch==0.28.1
more-itertools==9.1.0
numpy==1.25.1
openpyxl==3.1.2
pandas==2.0.3
passlib==1.7.4
phonenumbers==8.12.48
Pillow==10.0.0
psutil==5.9.5
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
pycryptodome==3.10.4
pydantic==2.0.2
pydantic_core==2.1.2
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python-ulid==1.1.0
pytz==2023.3
redis==4.6.0
redis-om==0.1.3
referencing==0.30.2
reportlab==4.0.4
requests==2.31.0
requests-file==1.5.1
rfc3986==1.5.0
rpds-py==0.9.2
rsa==4.9
s3transfer==0.6.1
sentry-sdk==1.29.0
sib-api-v3-sdk==7.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.18
sse-starlette==1.6.5
starlette==0.27.0
supertokens-python==0.12.9
tldextract==3.1.0
tqdm==4.65.0
twilio==7.9.1
types-pyOpenSSL==23.2.0.1
types-redis==4.6.0.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
wrapt==1.15.0
